+++ src/front/pages/Evento.jsx
@@ -1,5 +1,4 @@
 import { useNavigate } from "react-router-dom";
-import { VistaHome } from "./VistaHome";
 
 export const Evento = () => {
